Add files via upload
Gradient Descent is an iterative optimization algorithm used to minimize the difference between predicted and actual values in a regression problem. It works by updating model parameters, such as slope and intercept, in small steps proportional to the gradient of the cost function with respect to those parameters. By iteratively adjusting these parameters based on the direction that minimizes the cost, Gradient Descent gradually converges towards the optimal solution, aiming to find the best-fitting line for a given set of data points.