Add files via upload
Lemmatization is a linguistic process that reduces words to their base or root form, known as a lemma. Unlike stemming, which simply chops off prefixes or suffixes, lemmatization considers a word's meaning and context to produce accurate base forms. It's crucial in tasks like natural language processing for better understanding and analysis.